Medical_Panic_Detector/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
joblib>=1.1.0
pyserial>=3.5
pickle-mixin>=1.0.2
